myfunction/requirements.txt
requests
ansible-tower-cli